chore(): update istanbul-lib-instrument to support ESM when reading coverage
