+++ bricks/vs/src/code-editor/index.spec.tsx
@@ -7,7 +7,7 @@ import * as monaco from "monaco-editor/esm/vs/editor/editor.api.js";
 
 jest.mock("@next-core/theme", () => ({}));
 jest.mock("@next-core/react-runtime");
-jest.mock("./workers/yamlLinter", () => ({}));
+jest.mock("./workers/yamlLinter.js", () => ({}));
 
 global.ResizeObserver = ResizeObserver as any;
