+++ package.json
@@ -101,6 +101,7 @@
     "@babel/runtime": "^7.24.4",
     "@types/hast": "^3.0.4",
     "@types/react": "^18.2.79",
+    "istanbul-lib-instrument": "^6.0.3",
     "lodash": "^4.17.21",
     "moment": "^2.30.1"
   }
